Donut3DScreen: Refactor to separate 3D donut rendering logic from screen-specific UI and layout.
Modifies:
- TUI/vcxproj
- TUI/vcxproj.filters
-Screens/Donut3DScreen.h/.cpp

Adds:
- Screens/Content/Donut3D.h/.cpp

Key changes:

- Introduced Donut3D (Screens/Content/Donut3D.h/.cpp)
  - Encapsulates all donut simulation and rendering logic:
  - Renders into a caller-provided viewport (Rect)

- Simplified Donut3DScreen (Screens/Donut3DScreen.h/.cpp)
  - Owns all UI/chrome:
  - Defines viewport region via PageComposer and passes it to Donut3D::draw()

- Established viewport-based rendering contract
  - Donut3D no longer assumes full-screen rendering
  - Layout is fully controlled by the host (Donut3DScreen)

- Enables Donut3D reuse in:
  - windows
  - popups
  - split-screen layouts
  - future widget system
